examples/image_labeler_shapefile_downloader/requirement.txt
fiona
requests
datetime
rasterio
numpy
Pillow
